+++ Freechat/Bin/freechat.h
@@ -1,5 +1,5 @@
 /*
-***Copyright (C) 2019 Softwater, Inc
+***Copyleft (C) 2020 Softwater, Inc
 ***Contact: bogdyname@gmail.com
 */
 
@@ -8,9 +8,12 @@
 
 #include "Bin/bin.h"
 #include "ui_freechat.h"
+#include "Network/connectionf2f.h"
 #include <QTextTableFormat>
 #include <QMessageBox>
 #include <QtWidgets>
+#include <QMetaEnum>
+#include <QString>
 #include <QDialog>
 
 using namespace Qt;
@@ -20,46 +23,45 @@ class Freechat : public QDialog, private Ui::Freechat
     Q_OBJECT
 
 public:
-    static QString globalBuffer;
-    static QString viewField;
-
-public:
     static QString yourIp;
+    static QString command;
     static QString lanIpOfPeer;
     static QString wanIpOfPeer;
     static QString nickNameOfPeer;
     static QString bufferOfMessages;
+    static unsigned short int value;
+    QStringList commandsList;
 
-public:
-    Freechat(QWidget *parent = nullptr);
-    ~Freechat();
+    //UI
+    static QListWidget *listWithNickName;
+    static QTextEdit *viewField;
+    static QLineEdit *commandLine;
+    static QLineEdit *writeNickOfPeer;
+    static QLineEdit *writeLanIpOfPeer;
+    static QLineEdit *writeWanIpOfPeer;
+    static QLineEdit *lineForTypeText;
 
 private:
-    void AskForConnectingToPortPeer();
-    void AskForDisconnectingFromPortPeer();
-    bool ReplyFromPortPeer(bool &reply);
+    int (*checkNetworkConnection)() = nullptr;
+    bool (*checkConnection)() = nullptr;
 
-signals:
-    void SetTextInsideFiledOfChat();
+public:
+    explicit Freechat(QWidget *parent = nullptr);
+    ~Freechat();
 
 private slots:
-    void PassMessagesInsideBuffer();
+    void lineForTypeText_returnPressed();
+    void writeWanIpOfPeer_returnPressed();
+    void writeLanIpOfPeer_returnPressed();
+    void writeNickOfPeer_returnPressed();
+    void CommandLineInterface();
+    void networkInformation();
+    void connectionToPeerIn();
 
-    void on_showNetworkInfo_clicked(bool checked);
-    void on_connectionToPeer_clicked(bool checked);
+    void ServerStillWorking();// test code
 
-    void on_writeNickOfPeer_textChanged();
-    void on_writeLanIpOfPeer_textChanged();
-    void on_writeWanIpOfPeer_textChanged();
-    void on_wanButton_clicked(bool checked);
-    void on_lanButton_clicked(bool checked);
-    void on_nickButton_clicked(bool checked);
-
-    void on_listWithNickName_itemDoubleClicked(QListWidgetItem *item);
-
-    void on_lineForTypeText_textEdited(QString &messages);
-
-    void on_lineForTypeText_returnPressed();
+private:
+   QString status;
 
 private:
     Ui::Freechat *ui;
