+++ Freechat/Data/cryptography.cpp
@@ -1,0 +1,11 @@
+/*
+***Copyleft (C) 2020 Softwater, Inc
+***Contact: bogdyname@gmail.com
+*/
+
+#include "Data/cryptography.h"
+
+cryptography::cryptography()
+{
+
+}
