+++ Freechat/Data/cryptography.h
@@ -1,0 +1,14 @@
+/*
+***Copyleft (C) 2020 Softwater, Inc
+***Contact: bogdyname@gmail.com
+*/
+
+#ifndef CRYPTOGRAPHY_H
+#define CRYPTOGRAPHY_H
+
+class cryptography
+{
+public:
+    cryptography();
+};
+#endif
