+++ Freechat/Data/datasave.h
@@ -1,5 +1,5 @@
 /*
-***Copyright (C) 2019 Softwater, Inc
+***Copyleft (C) 2020 Softwater, Inc
 ***Contact: bogdyname@gmail.com
 */
 
@@ -19,31 +19,7 @@ private ConnectionF2F, private Freechat
     Q_OBJECT
 
 public:
-    QString *bufferNickname = nullptr;
-
-private:
-    QByteArray *buffer = nullptr;
-    QStorageInfo storage = QStorageInfo::root();
-
-public:
+    Datasave();
     ~Datasave();
-    Datasave(Freechat *parent = nullptr);
-    Datasave(QFile &fileWithData, QFile &fileWithDataForBackup);
-
-signals:
-    void CheckYourMemorySize();
-    void ChooseFileWithData(QFile &fileWithData);
-    void ReadFileForViewMessages(QFile &file, QString &nickname);
-
-private slots:
-    inline void DeleteAllDataForFreeMemory(QFile &fileWithData, QFile &fileWithDataForBackup);
-    void RunBackupFiles(QFile &fileWithData, QFile &fileWithDataForBackup);
-    inline void ReadFile(QFile &fileWithData);
-    void OpenFile(QFile &fileWithData);
-
-private:
-    void RSAMODULE(QFile &fileWithData);
-    inline bool CheckNicknameForSaveFile(QString &nickname);
-    inline QString ReadFirstStringFromDataFile(QString &nickname);
 };
 #endif
