+++ Freechat/Network/peerin.h
@@ -1,17 +1,18 @@
 /*
-***Copyright (C) 2019 Softwater, Inc
+***Copyleft (C) 2020 Softwater, Inc
 ***Contact: bogdyname@gmail.com
 */
 
 #ifndef PEERIN_H
 #define PEERIN_H
 
-#include "Network/connectionf2f.h"
-#include "Network/peerout.h"
+#include "Bin/freechat.h"
+#include <QTcpServer>
+#include <QTcpSocket>
 
-class Peerout;
-class QTcpSpcket;
-class ConnectionF2F;
+class Freechat;
+class QTcpSocket;
+class QTcpServer;
 
 class Peerin : public QTcpServer
 {
@@ -19,18 +20,18 @@ class Peerin : public QTcpServer
 
 private:
     QTcpServer *server = nullptr;
-    quint16 nextBlockSize = 0;
+    quint16 nextBlockSize;
+    QTcpSocket *socket = nullptr;
 
-private:
+public:
     explicit Peerin(QObject *parent = nullptr);
     ~Peerin();
 
-private:
-    void SendResponseToClient(QTcpSocket *socket, QString &messages);
-
 public slots:
     virtual void SlotNewConnection();
-    void SlotReadClient ();
+    void SendResponseToClient();
+    void SlotReadClient();
+    void clearValue();
 
 };
 #endif
