+++ Freechat/main.cpp
@@ -1,25 +1,22 @@
 /*
-***Copyright (C) 2019 Softwater, Inc
+***Copyleft (C) 2020 Softwater, Inc
 ***Contact: bogdyname@gmail.com
 */
 
 #include "Bin/freechat.h"
-#include "Data/datasave.h"
-#include <QtGui>
 #include <QApplication>
-#include <QtCore/QSettings>
-#include <QtNetwork/QNetworkSession>
-#include <QtNetwork/QNetworkConfigurationManager>
-
-using namespace Qt;
 
 int main(int argc, char *argv[])
 {
     QApplication app(argc, argv);
 
-    QNetworkConfigurationManager manager;
+    QFile file(":/source/style.css");
+    file.open(QFile::ReadOnly);
+    app.setStyleSheet(file.readAll());
 
-    Freechat freechat;
-    freechat.show();
+    // ui
+    Freechat window;
+    window.show();
+
     return app.exec();
 }
