+++ Freechat/resources.qrc
