+++ Freechat/test.pro
@@ -1,14 +1,7 @@
-#---------------------------------
-#Freechat (decentralized chat)
-#Freechat via C++ & Qt
-#Debugger(Qt 5.12.0 for UWP 64bit (MSVC 2015)2)
-#Copyright (C) 2019 Softwater, Inc.
-#---------------------------------
-
 QT       += gui
-QT       += svg
 QT       += network
 QT       += core gui
+QT       += widgets
 
 greaterThan(QT_MAJOR_VERSION, 5): QT += widgets
 
@@ -20,30 +13,32 @@ DEFINES += QT_DEPRECATED_WARNINGS
 CONFIG += c++11
 CONFIG += c++14
 CONFIG += c++17
-CONFIG += qt
 
 SOURCES += \
         Bin/bin.cpp \
+        Bin/freechat.cpp \
+        Data/cryptography.cpp \
         Network/connectionf2f.cpp \
-        Network/cryptography.cpp \
         Network/peerin.cpp \
         Network/peerout.cpp \
-        main.cpp \
-        Bin/freechat.cpp \
-        Data/datasave.cpp
+        Data/datasave.cpp \
+        main.cpp
 
 HEADERS += \
         Bin/bin.h \
+        Data/cryptography.h \
         Network/connectionf2f.h \
         Bin/freechat.h \
-        Data/datasave.h \
-        Network/cryptography.h \
         Network/peerin.h \
-        Network/peerout.h
+        Network/peerout.h \
+        Data/datasave.h
 
 FORMS += \
-        Designer/freechat.ui
+        freechat.ui
 
 qnx: target.path = /tmp/$${TARGET}/bin
 else: unix:!android: target.path = /opt/$${TARGET}/bin
 !isEmpty(target.path): INSTALLS += target
+
+RESOURCES += \
+    resources.qrc
