--- build-test-Desktop_Qt_5_12_0_MSVC2017_64bit2-Debug/debug/qrc_resources.cpp
@@ -1,131 +1,0 @@
-/****************************************************************************
-** Resource object code
-**
-** Created by: The Resource Compiler for Qt version 5.12.0
-**
-** WARNING! All changes made in this file will be lost!
-*****************************************************************************/
-
-static const unsigned char qt_resource_data[] = {
-  // D:/Freechat-Windows/test/source/style.css
-  0x0,0x0,0x2,0xad,
-  0x51,
-  0x57,0x69,0x64,0x67,0x65,0x74,0xd,0xa,0x7b,0xd,0xa,0x20,0x20,0x20,0x20,0x66,
-  0x6f,0x6e,0x74,0x3a,0x20,0x38,0x70,0x74,0x20,0x22,0x4c,0x65,0x74,0x74,0x65,0x72,
-  0x20,0x47,0x6f,0x74,0x68,0x69,0x63,0x20,0x53,0x74,0x64,0x22,0x2c,0x20,0x77,0x68,
-  0x69,0x74,0x65,0x3b,0xd,0xa,0x20,0x20,0x20,0x20,0x62,0x61,0x63,0x6b,0x67,0x72,
-  0x6f,0x75,0x6e,0x64,0x2d,0x63,0x6f,0x6c,0x6f,0x72,0x3a,0x20,0x62,0x6c,0x61,0x63,
-  0x6b,0x3b,0xd,0xa,0x20,0x20,0x20,0x20,0x63,0x6f,0x6c,0x6f,0x72,0x3a,0x20,0x23,
-  0x46,0x46,0x44,0x37,0x30,0x30,0x3b,0xd,0xa,0x7d,0xd,0xa,0xd,0xa,0x51,0x54,
-  0x65,0x78,0x74,0x45,0x64,0x69,0x74,0xd,0xa,0x7b,0xd,0xa,0x20,0x20,0x20,0x20,
-  0x62,0x6f,0x72,0x64,0x65,0x72,0x3a,0x20,0x31,0x70,0x78,0x20,0x73,0x6f,0x6c,0x69,
-  0x64,0x20,0x23,0x30,0x30,0x46,0x46,0x46,0x46,0x3b,0xd,0xa,0x20,0x20,0x20,0x20,
-  0x62,0x61,0x63,0x6b,0x67,0x72,0x6f,0x75,0x6e,0x64,0x2d,0x63,0x6f,0x6c,0x6f,0x72,
-  0x3a,0x20,0x62,0x6c,0x61,0x63,0x6b,0x3b,0xd,0xa,0x20,0x20,0x20,0x20,0x63,0x6f,
-  0x6c,0x6f,0x72,0x3a,0x20,0x23,0x46,0x46,0x44,0x37,0x30,0x30,0x3b,0xd,0xa,0x7d,
-  0xd,0xa,0xd,0xa,0x51,0x4c,0x69,0x73,0x74,0x57,0x69,0x64,0x67,0x65,0x74,0xd,
-  0xa,0x7b,0xd,0xa,0x20,0x20,0x20,0x20,0x62,0x6f,0x72,0x64,0x65,0x72,0x3a,0x20,
-  0x31,0x70,0x78,0x20,0x73,0x6f,0x6c,0x69,0x64,0x20,0x23,0x30,0x30,0x46,0x46,0x46,
-  0x46,0x3b,0xd,0xa,0x20,0x20,0x20,0x20,0x62,0x61,0x63,0x6b,0x67,0x72,0x6f,0x75,
-  0x6e,0x64,0x2d,0x63,0x6f,0x6c,0x6f,0x72,0x3a,0x20,0x62,0x6c,0x61,0x63,0x6b,0x3b,
-  0xd,0xa,0x20,0x20,0x20,0x20,0x63,0x6f,0x6c,0x6f,0x72,0x3a,0x20,0x23,0x46,0x46,
-  0x44,0x37,0x30,0x30,0x3b,0xd,0xa,0x7d,0xd,0xa,0xd,0xa,0xd,0xa,0x51,0x4c,
-  0x69,0x6e,0x65,0x45,0x64,0x69,0x74,0xd,0xa,0x7b,0xd,0xa,0x20,0x20,0x20,0x20,
-  0x62,0x6f,0x72,0x64,0x65,0x72,0x3a,0x20,0x31,0x70,0x78,0x20,0x73,0x6f,0x6c,0x69,
-  0x64,0x20,0x23,0x30,0x30,0x46,0x46,0x46,0x46,0x3b,0xd,0xa,0x20,0x20,0x20,0x20,
-  0x62,0x6f,0x72,0x64,0x65,0x72,0x2d,0x72,0x61,0x64,0x69,0x75,0x73,0x3a,0x20,0x35,
-  0x70,0x78,0x3b,0xd,0xa,0x20,0x20,0x20,0x20,0x70,0x61,0x64,0x64,0x69,0x6e,0x67,
-  0x3a,0x20,0x30,0x20,0x38,0x70,0x78,0x3b,0xd,0xa,0x7d,0xd,0xa,0xd,0xa,0x51,
-  0x4c,0x61,0x62,0x65,0x6c,0xd,0xa,0x7b,0xd,0xa,0x20,0x20,0x20,0x20,0x62,0x6f,
-  0x72,0x64,0x65,0x72,0x3a,0x20,0x31,0x70,0x78,0x20,0x73,0x6f,0x6c,0x69,0x64,0x20,
-  0x23,0x30,0x30,0x46,0x46,0x46,0x46,0x3b,0xd,0xa,0x20,0x20,0x20,0x20,0x62,0x6f,
-  0x72,0x64,0x65,0x72,0x2d,0x72,0x61,0x64,0x69,0x75,0x73,0x3a,0x20,0x35,0x70,0x78,
-  0x3b,0xd,0xa,0x20,0x20,0x20,0x20,0x70,0x61,0x64,0x64,0x69,0x6e,0x67,0x3a,0x20,
-  0x32,0x70,0x78,0x3b,0xd,0xa,0x20,0x20,0x20,0x20,0x63,0x6f,0x6c,0x6f,0x72,0x3a,
-  0x20,0x23,0x46,0x46,0x44,0x37,0x30,0x30,0x3b,0xd,0xa,0x20,0x20,0x20,0x20,0x66,
-  0x6f,0x6e,0x74,0x3a,0x20,0x38,0x70,0x74,0x20,0x22,0x4c,0x65,0x74,0x74,0x65,0x72,
-  0x20,0x47,0x6f,0x74,0x68,0x69,0x63,0x20,0x53,0x74,0x64,0x22,0x3b,0xd,0xa,0x7d,
-  0xd,0xa,0xd,0xa,0x51,0x50,0x75,0x73,0x68,0x42,0x75,0x74,0x74,0x6f,0x6e,0xd,
-  0xa,0x7b,0xd,0xa,0x20,0x20,0x20,0x20,0x62,0x6f,0x72,0x64,0x65,0x72,0x3a,0x20,
-  0x31,0x2e,0x35,0x70,0x78,0x20,0x73,0x6f,0x6c,0x69,0x64,0x20,0x23,0x30,0x30,0x46,
-  0x46,0x46,0x46,0x3b,0xd,0xa,0x20,0x20,0x20,0x20,0x62,0x6f,0x72,0x64,0x65,0x72,
-  0x2d,0x72,0x61,0x64,0x69,0x75,0x73,0x3a,0x20,0x35,0x70,0x78,0x3b,0xd,0xa,0x20,
-  0x20,0x20,0x20,0x63,0x6f,0x6c,0x6f,0x72,0x3a,0x20,0x23,0x46,0x46,0x44,0x37,0x30,
-  0x30,0x3b,0xd,0xa,0x20,0x20,0x20,0x20,0x70,0x61,0x64,0x64,0x69,0x6e,0x67,0x3a,
-  0x20,0x30,0x20,0x38,0x70,0x78,0x3b,0xd,0xa,0x7d,0xd,0xa,
-  
-};
-
-static const unsigned char qt_resource_name[] = {
-  // source
-  0x0,0x6,
-  0x7,0xa6,0xc8,0x95,
-  0x0,0x73,
-  0x0,0x6f,0x0,0x75,0x0,0x72,0x0,0x63,0x0,0x65,
-    // style.css
-  0x0,0x9,
-  0x0,0x28,0xbf,0x23,
-  0x0,0x73,
-  0x0,0x74,0x0,0x79,0x0,0x6c,0x0,0x65,0x0,0x2e,0x0,0x63,0x0,0x73,0x0,0x73,
-  
-};
-
-static const unsigned char qt_resource_struct[] = {
-  // :
-  0x0,0x0,0x0,0x0,0x0,0x2,0x0,0x0,0x0,0x1,0x0,0x0,0x0,0x1,
-0x0,0x0,0x0,0x0,0x0,0x0,0x0,0x0,
-  // :/source
-  0x0,0x0,0x0,0x0,0x0,0x2,0x0,0x0,0x0,0x1,0x0,0x0,0x0,0x2,
-0x0,0x0,0x0,0x0,0x0,0x0,0x0,0x0,
-  // :/source/style.css
-  0x0,0x0,0x0,0x12,0x0,0x0,0x0,0x0,0x0,0x1,0x0,0x0,0x0,0x0,
-0x0,0x0,0x1,0x6f,0xff,0xf7,0xea,0x78,
-
-};
-
-#ifdef QT_NAMESPACE
-#  define QT_RCC_PREPEND_NAMESPACE(name) ::QT_NAMESPACE::name
-#  define QT_RCC_MANGLE_NAMESPACE0(x) x
-#  define QT_RCC_MANGLE_NAMESPACE1(a, b) a##_##b
-#  define QT_RCC_MANGLE_NAMESPACE2(a, b) QT_RCC_MANGLE_NAMESPACE1(a,b)
-#  define QT_RCC_MANGLE_NAMESPACE(name) QT_RCC_MANGLE_NAMESPACE2( \
-        QT_RCC_MANGLE_NAMESPACE0(name), QT_RCC_MANGLE_NAMESPACE0(QT_NAMESPACE))
-#else
-#   define QT_RCC_PREPEND_NAMESPACE(name) name
-#   define QT_RCC_MANGLE_NAMESPACE(name) name
-#endif
-
-#ifdef QT_NAMESPACE
-namespace QT_NAMESPACE {
-#endif
-
-bool qRegisterResourceData(int, const unsigned char *, const unsigned char *, const unsigned char *);
-
-bool qUnregisterResourceData(int, const unsigned char *, const unsigned char *, const unsigned char *);
-
-#ifdef QT_NAMESPACE
-}
-#endif
-
-int QT_RCC_MANGLE_NAMESPACE(qInitResources_resources)();
-int QT_RCC_MANGLE_NAMESPACE(qInitResources_resources)()
-{
-    QT_RCC_PREPEND_NAMESPACE(qRegisterResourceData)
-        (0x2, qt_resource_struct, qt_resource_name, qt_resource_data);
-    return 1;
-}
-
-int QT_RCC_MANGLE_NAMESPACE(qCleanupResources_resources)();
-int QT_RCC_MANGLE_NAMESPACE(qCleanupResources_resources)()
-{
-    QT_RCC_PREPEND_NAMESPACE(qUnregisterResourceData)
-       (0x2, qt_resource_struct, qt_resource_name, qt_resource_data);
-    return 1;
-}
-
-namespace {
-   struct initializer {
-       initializer() { QT_RCC_MANGLE_NAMESPACE(qInitResources_resources)(); }
-       ~initializer() { QT_RCC_MANGLE_NAMESPACE(qCleanupResources_resources)(); }
-   } dummy;
-}
