--- test/Network/cryptography.cpp
@@ -1,11 +1,0 @@
-/*
-***Copyleft (C) 2019 Softwater, Inc
-***Contact: bogdyname@gmail.com
-*/
-
-#include "Network/cryptography.h"
-
-Cryptography::Cryptography()
-{
-
-}
