--- test/Network/peerout.h
@@ -1,37 +1,0 @@
-/*
-***Copyright (C) 2019 Softwater, Inc
-***Contact: bogdyname@gmail.com
-*/
-
-#ifndef PEEROUT
-#define PEEROUT
-
-#include "Bin/freechat.h"
-#include <QHostAddress>
-#include <QTcpSocket>
-
-
-class Freechat;
-class QTcpSocket;
-class ConnectionF2F;
-
-class Peerout : public QTcpSocket
-{
-    Q_OBJECT
-
-public:
-    explicit Peerout();
-    ~Peerout();
-
-public slots:
-    void SlotReadyRead();
-    void SlotConnected();
-    void SlotConnecting();
-    void SlotSendToServer();
-    void SlotError(QAbstractSocket::SocketError);
-
-private:
-    QTcpSocket *socket;
-    quint16 nextBlockSize;
-};
-#endif
